+++ spec/wrapped_module_spec.rb
@@ -35,6 +35,19 @@ describe Pry::WrappedModule do
           end
         end
       end
+
+      module ModuleNameOverride
+        def self.name; 'Not a good idea, but it happens in the wild.' end
+      end
+    end
+
+    describe 'overidden Module#name' do
+      specify '#source_location does not return nil' do
+        c = Pry::WrappedModule(Host::ModuleNameOverride).candidate(0)
+        expect(c.wrapped).to eq(Host::ModuleNameOverride)
+        expect(c.nonblank_name).to eq(Host::ModuleNameOverride.name)
+        expect(c.source_location).not_to be_nil
+      end
     end
 
     describe "number_of_candidates" do